Fix file locking on Mono (#109)
Mono behaves differently to dotnet and .NET on Windows with file
locks. Creating a FileStream with FileShare.None will not lock the
file when running on Mono. For this to work on Mono FileStream.Lock
has to be called.

Added Mono platform detection and the CrossPlatLock will call
FileStream.Lock on Mono. Also using FileOptions.DeleteOnClose is
a problem on Mono since the file can be deleted from another process
without a lock so the lock file is not deleted when on Mono.